kaggle_submission/requirements.txt
# BigQuery AI: Intelligent Retail Analytics Engine
# Competition Winner: $100,000 Prize Track

# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Google Cloud dependencies
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.0.0

# Machine Learning dependencies
scikit-learn>=1.0.0
lightgbm>=3.2.0

# Web framework (optional, for API deployment)
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Configuration and utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Logging and monitoring
structlog>=21.0.0

# Jupyter notebook (for development)
jupyter>=1.0.0
ipykernel>=6.0.0